Update MANIFEST.SKIP for Clownfish.
Building the Clownfish compiler is a prerequisite for generating POD files
needed by the "dist" action which builds the CPAN distro for Lucy, but it
leaves a lot of detritus lying around.  This commit edits MANIFEST.SKIP so
that the generated files are not included in the Lucy CPAN tarball.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1098908 13f79535-47bb-0310-9956-ffa450edef68