Strip Charmonizer libs and flags.
Don't include assorted guilibs and winsocks when compiling Charmonizer itself.
Charmonizer itself has a very simple C89 code base, and is designed to be a
standalone app.  (Charmonizer's *tests* on the other hand, *must* be compiled with
the same compiler flags that get passed to charmonize.exe when it is used to 
generate charmony.h.)

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1129408 13f79535-47bb-0310-9956-ffa450edef68