Improve Windows "clean" target for Charmonizer.
The DEL builtin in cmd.exe returns failure when any one of the file arguments
passed to it  do not exist.  NMake ignores this failure, but DMake aborts.  To
solve the problem, use a FOR loop and an IF EXIST test so that DEL is only
invoked for files that exist.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1139581 13f79535-47bb-0310-9956-ffa450edef68