Insert sanity checks into string processing routines.
Some string processing routines may behave incorrectly when operating on data
which is not well-formed UTF-8.  They are never supposed to see ill-formed
data, but if there is a bug in either Lucy or in an application feeding Lucy,
the possibility exists.

As part of a defense in depth against possible security holes, these routines
should abort when they detect invalid data.  This will turn flaws which
potentially could worsen into severe problems such as buffer overflow
vulnerabilities into at worst denial-of-service vulnerabilities.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1165380 13f79535-47bb-0310-9956-ffa450edef68