Unify IDENTIFIER and CLASS_NAME_COMPONENT tokens.
Bottom-up parsers such as the LALR(1) parsers built with Lemon or YACC have
more difficulty resolving grammar ambiguities than recursive descent parsers,
because they operate with less context.  Unless we want to restrict IDENTIFIER
so that it can never be part of a class name (for instance by making it all
lower case), we end up with conflicts where it is not clear how a token should
be reduced.  Therefore we should tolerate class names and object type names
which are invalid according to the rules of Clownfish at the parser grammar
level, but catch the invalid names later at object construction time.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/branches/clownfish_lemon@1187919 13f79535-47bb-0310-9956-ffa450edef68