Use memory pool instead of malloc/free in parser.
Use a memory pool for allocating all C strings inside both the Flex-based
scanner and the Lemon-based parser, deallocating them en masse by destroying
the memory pool once the parse operation completes..  This is a natural use
for a memory pool because each parse() invocation is a substantial
undertaking, yet also contained and monolithic.  The primary benefit is code
simplification (because we no longer have to worry about freeing C strings
inside the lemon grammar).  The secondary benefit is that it is harder to leak
memory, especially when a parse operation fails.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/branches/clownfish_lemon@1188972 13f79535-47bb-0310-9956-ffa450edef68