Capture Flex-generated files in svn.
Flex will not be present on every system, and unlike Lemon, it's too big to
bundle.  Therefore, store its generated files in svn, where they only need
change if the upstream .l file changes.

Flex is under a BSD license; these files were generated by version 2.5.35.
The COPYING file in the Flex 2.5.35 source code includes the following
passage:

    Note that the "flex.skl" scanner skeleton carries no copyright notice.
    You are free to do whatever you please with scanners generated using flex;
    for them, you are not even bound by the above copyright.

Therefore, it is not necessary to modify either LICENSE or NOTICE to
accommodate these files.  They do not need (and should not have) ASF license
headers.  The only additional action we need to take is to modify the
rat-excludes file.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/branches/clownfish_lemon@1189573 13f79535-47bb-0310-9956-ffa450edef68