Implement CaseFolder by means of Normalizer
This commit removes all code of CaseFolder specific to host languages and
reimplements the CaseFolder transform using a Lucy::Analysis::Normalizer for
backward compatibility.

The CaseFolder is now deprecated.

There are differences in the mapping of a few Unicode characters compared to
the old Perl implementation, so the changes are not 100% backward compatible.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1243099 13f79535-47bb-0310-9956-ffa450edef68