Blacklist methods which shouldn't be bound.
Change from whitelisting methods to be bound to blacklisting methods which
should be not be bound.  This increases the size of the compiled object
somewhat, but lowers maintenance costs by decreasing the amount of code it
takes to maintain a host binding and by automatically binding new methods as
they are added.

Some methods can only be bound manually because they contain types which
cannot be mapped automatically.  These are now silently excluded from binding
by default.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1244363 13f79535-47bb-0310-9956-ffa450edef68